requirements.txt
aiohttp>=3.9,<4
requests>=2.31,<3
python-dotenv>=1.0,<2
pandas>=2.0,<3
pytest>=7.0,<8
pytest-asyncio>=0.21,<1
pytest-aiohttp>=1.0,<2

# Caching for aiohttp handlers
aiohttp-cache>=2.2,<3


